% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attention.R
\name{attention_heads}
\alias{attention_heads}
\title{Get attention heads}
\usage{
attention_heads(model, dataset, n = 100)
}
\arguments{
\item{model}{the tabtransformer model.}

\item{dataset}{dataset to pass through the model to generate attention heads.}

\item{n}{number of rows to use from the dataset. Default is 100.}
}
\value{
average attention heads as a matrix
}
\description{
Get attention heads
}
