% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blastchar}
\alias{blastchar}
\title{Customer churn dataset from Kaggle.}
\format{
A data frame with 7,043 rows and 21 variables:
\describe{
  \item{income}{income as a binary variable}
  ...
}
}
\source{
\url{https://www.kaggle.com/blastchar/telco-customer-churn}
}
\usage{
blastchar
}
\description{
Each row represents a customer, each column contains customer’s attributes.
}
\keyword{datasets}
