% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Income dataset from Kaggle.}
\format{
A data frame with 32,561 rows and 15 variables:
\describe{
  \item{income}{income as a binary variable}
  ...
}
}
\source{
\url{https://www.kaggle.com/lodetomasi1995/income-classification}
}
\usage{
income
}
\description{
A dataset containing incomes (<=50K vs. >50K) and predictors.
A common becnhmark dataset for tabular models.
}
\keyword{datasets}
