% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attention.R
\name{intersample_attention_heads}
\alias{intersample_attention_heads}
\title{Get intersample attention heads}
\usage{
intersample_attention_heads(model, dataset, n = 100, device = "cpu")
}
\arguments{
\item{model}{the tabtransformer model.}

\item{dataset}{dataset to pass through the model to generate intersample
attention heads.}

\item{n}{number of rows to use from the dataset. Default is 100.}

\item{device}{whether to run this on 'cpu' or 'cuda'}
}
\value{
average intersample attention heads as a matrix
}
\description{
Get intersample attention heads
}
