% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_callback.R
\name{neptune_callback}
\alias{neptune_callback}
\title{Neptune Callback}
\usage{
neptune_callback(
  project_name = NULL,
  name = "",
  tags = NULL,
  params = NULL,
  api_token = "ANONYMOUS"
)
}
\arguments{
\item{project_name}{name of the project}

\item{name}{name of the run}

\item{tags}{vector of tags}

\item{params}{vector of hyperparameter names to log}

\item{api_token}{your Neptune API token}
}
\value{
logs metrics to Neptune
}
\description{
Neptune Callback
}
