% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_model.R
\name{tab_model}
\alias{tab_model}
\title{Tabular Model.}
\usage{
tab_model(x, model, ...)
}
\arguments{
\item{x}{Either a tidymodels recipe or a data.frame}

\item{model}{model type to be used - default is a transformer type model.}

\item{...}{hyperparameters passed to the model specification}
}
\value{

}
\description{
Tabular Model.
}
