% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_dataset.R
\name{tabular_dataset.data.frame}
\alias{tabular_dataset.data.frame}
\title{Tabular Dataset with data.frame input.}
\usage{
\method{tabular_dataset}{data.frame}(x, response, cat_vars = NULL, cont_vars = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{response}{response variable name (only used if x is a data.frame)}

\item{cat_vars}{vector of categorical variable names (only used if x is a data.frame)}

\item{cont_vars}{vector of continuous variable names (only used if x is a data.frame)}
}
\value{

}
\description{
Tabular Dataset with data.frame input.
}
