% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tabular_dataset}
\alias{tabular_dataset}
\title{Tabular Dataset.}
\usage{
tabular_dataset(x, response, cat_vars, cont_vars)
}
\arguments{
\item{x}{either a recipe from the {recipes} package or a data.frame}

\item{response}{response variable name (only used if x is a data.frame)}

\item{cat_vars}{vector of categorical variable names (only used if x is a data.frame)}

\item{cont_vars}{vector of continuous variable names (only used if x is a data.frame)}
}
\value{

}
\description{
Tabular Dataset.
}
