% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_dataset.R
\name{tabular_dataset.recipe}
\alias{tabular_dataset.recipe}
\title{Tabular Dataset with recipe input.}
\usage{
\method{tabular_dataset}{recipe}(x, data)
}
\arguments{
\item{x}{a recipe}

\item{data}{a dataset to be prepped by the recipe}
}
\value{

}
\description{
Tabular Dataset with recipe input.
}
