% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_callback.R
\name{var_importance_callback}
\alias{var_importance_callback}
\title{Importance Callback}
\usage{
var_importance_callback(
  names,
  top = 10,
  start_epoch = 1,
  retain_data = TRUE,
  retain_plot = TRUE
)
}
\arguments{
\item{names}{variable names as they appear in the model input x}

\item{top}{number of variables to plot (top n). Default is 10.}

\item{start_epoch}{the epoch to start plotting from. Default is 1.}
}
\value{
two outputs saved to the model context (importance_plot and importance_data)
}
\description{
Importance Callback
}
