% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attention.R
\name{intersample_attention_heads}
\alias{intersample_attention_heads}
\title{Get intersample attention heads}
\usage{
intersample_attention_heads(model, dataset, n = 100, cut = 0.99)
}
\arguments{
\item{model}{the tabtransformer model.}

\item{dataset}{dataset to pass through the model to generate intersample
attention heads.}

\item{n}{number of rows to use from the dataset. Default is 100.}

\item{cut}{quantile below which attention is rounded down to 0. Default is
0.99 (i.e. only retain the top 1% of attention heads).}
}
\value{
average intersample attention heads as a matrix
}
\description{
Get intersample attention heads
}
